fix previous, move `Amount` and `TokenId` to `mc-transaction-types`
