+++ Cargo.toml
@@ -123,6 +123,7 @@ members = [
     "transaction/core",
     "transaction/core/test-utils",
     "transaction/std",
+    "transaction/types",
     "util/b58-decoder",
     "util/build/enclave",
     "util/build/grpc",
