+++ crypto/ring-signature/signer/Cargo.toml
@@ -3,6 +3,7 @@ name = "mc-crypto-ring-signature-signer"
 version = "1.3.0-pre0"
 authors = ["MobileCoin"]
 edition = "2021"
+readme = "README.md"
 
 [dependencies]
 # External dependencies
@@ -19,6 +20,7 @@ zeroize = { version = "1", default-features = false }
 mc-account-keys = { path = "../../../account-keys", default-features = false }
 mc-crypto-keys = { path = "../../keys", default-features = false }
 mc-crypto-ring-signature = { path = "..", default-features = false }
+mc-transaction-types = { path = "../../../transaction/types" }
 mc-util-serial = { path = "../../../util/serial" }
 
 [target.'cfg(any(target_feature = "avx2", target_feature = "avx"))'.dependencies]
