+++ transaction/core/tests/digest-test-vectors.rs
@@ -1,3 +1,5 @@
+// Copyright (c) 2018-2022 The MobileCoin Foundation
+
 use mc_account_keys::AccountKey;
 use mc_crypto_digestible_test_utils::*;
 use mc_crypto_keys::RistrettoPrivate;
@@ -6,8 +8,7 @@ use mc_transaction_core::{
     BlockVersion, Token,
 };
 use mc_util_from_random::FromRandom;
-use rand_core::{RngCore, SeedableRng};
-use rand_hc::Hc128Rng as FixedRng;
+use mc_util_test_helper::{RngCore, RngType as FixedRng, SeedableRng};
 
 fn test_accounts() -> Vec<AccountKey> {
     let mut rng: FixedRng = SeedableRng::from_seed([12u8; 32]);
